+++ Projects/App/Sources/Presentation/Calendar/View/CalendarRecordView.swift
@@ -7,6 +7,7 @@
 //
 
 import SwiftUI
+import DesignSystem
 
 enum RecordViewText {
     case beforeRecord
