+++ Projects/App/Sources/Presentation/Calendar/View/DatePicker.swift
@@ -8,6 +8,7 @@
 
 import SwiftUI
 import UIKit
+import DesignSystem
 
 struct DatePicker: View {
     @Binding var selectedYear: Int
