+++ Projects/App/Sources/Presentation/Login/View/LoginQualificationView.swift
@@ -7,6 +7,7 @@
 //
 
 import SwiftUI
+import DesignSystem
 
 enum Qualification: String {
     case personal = "https://ballistic-dollar-2f4.notion.site/820097f01cca4e929e189f925548071b" // 개인정보 처리방침
