+++ Projects/DesignSystem/Sources/DesignSystem/Color/ColorDemo.swift
@@ -1,0 +1,21 @@
+//
+//  ColorDemo.swift
+//  DesignSystem
+//
+//  Created by 박서연 on 2024/03/21.
+//  Copyright © 2024 Mogong. All rights reserved.
+//
+
+import SwiftUI
+
+struct ColorDemo: View {
+    var body: some View {
+        Text("gray 1")
+            .padding()
+            .background(Color.test)
+    }
+}
+
+#Preview {
+    ColorDemo()
+}
